Removed dependencies + Added installation information
Eliminated sarcastic dependencies on mouse and keyboard, streamlining the application's functionality. Introduced two installation methods for user convenience: 1) Navigate to the Releases tab and download the latest `.exe` file, and 2) Click on the "latest release" text to access and download the corresponding `.exe` file. This provides users with multiple options for a seamless installation experience.